+++ packages/react-magma-dom/src/components/Alert/Alert.stories.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Alert } from '.';
+
+export default {
+  component: Alert,
+  title: 'Alert',
+};
+
+export const Default = () => {
+  return (
+    <Alert isDismissible toastDuration={5000}>
+      Content Is Fun
+    </Alert>
+  );
+};