+++ packages/react-magma-dom/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### 2.3.5-alpha.1 (2021-01-26)
+### 2.3.5-alpha.2 (2021-01-29)
 
 **Note:** Version bump only for package react-magma-dom
 
@@ -14,6 +14,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 **Note:** Version bump only for package react-magma-dom
 
 =======
+
+### 2.3.5-alpha.1 (2021-01-26)
+
+**Note:** Version bump only for package react-magma-dom
 
 ### 2.3.5-alpha.0 (2021-01-26)