+++ packages/react-magma-landing/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-### 1.0.4-alpha.1 (2021-01-26)
+### 1.0.4-alpha.2 (2021-01-29)
 
 **Note:** Version bump only for package react-magma-landing
 
@@ -14,6 +14,10 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 **Note:** Version bump only for package react-magma-landing
 
 =======
+
+### 1.0.4-alpha.1 (2021-01-26)
+
+**Note:** Version bump only for package react-magma-landing
 
 ### 1.0.4-alpha.0 (2021-01-26)